backend:refactor: split linklinkuser views
